weighted_quantiles: move into its own module
